requirements.txt
alembic==1.0.11
certifi==2019.6.16
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.3.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-spec-validator==0.2.7
pudb==2018.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
swagger-ui-bundle==0.0.5
urllib3==1.25.3
urwid==2.0.1
Werkzeug==0.14.1
WTForms==2.2.1
